+++ drizzle/0012_easy_bucky.sql
@@ -1,0 +1,3 @@
+ALTER TABLE `song` ADD `notes_markdown` text DEFAULT '' NOT NULL;--> statement-breakpoint
+ALTER TABLE `song` ADD `notes_hash` text;--> statement-breakpoint
+ALTER TABLE `song` ADD `notes_version` integer DEFAULT 0 NOT NULL;