+++ src/lib/components/GlobalFooter.svelte
@@ -1,13 +1,18 @@
 <footer
-	class="page-x-padding pt-4 pb-4 border-t border-white/10 text-11px flex flex-wrap gap-x-6 gap-y-2"
+	class="page-x-padding pt-4 pb-4 border-t border-white/10 text-11px flex flex-wrap gap-x-6 gap-y-2 justify-between"
 >
 	<!-- <span>Stem Shovel</span> -->
 	<div class="opacity-70">
-		An open source experiment from <a
+		An experiment from <a
 			class="inline hover-underline underline-offset-4 hover-text-maximumYellow"
 			href="https://www.lightningjar.com">⚡️ Lightning Jar</a
 		>
 	</div>
 
-	<!-- <a class="underline underline-offset-4 hover:text-maximumYellow" href="/test">Static test page</a> -->
+	<!-- Rick Roll Easter Egg -->
+	<a
+		class="underline underline-offset-4 hover:text-maximumYellow opacity-70 underline underline-offset-4 hover-text-maximumYellow"
+		href="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
+		title="Click in case of emergency">For Your Eyes Only</a
+	>
 </footer>